Modified zulu JRE-FX installation
Modified documentation for the "Download and installation" section under the "With zulu JRE-FX:" subsection. The jar file is no longer named mcaselector-1.16.2.jar so following the original directions yields a "Unable to access jarfile mcaselector-1.16.2.jar" error. The jar file downloaded is named mcaselector.jar and having users drag and drop will be easier for the future.