add appstream file for linux
This file is a universal metadata that describe the application for Linux
software centers and package/catalog systems. Required to publish the
application in the Flathub repository (https://flathub.org).

Note that new releases should be defined under the `releases` tag, see:
https://docs.flathub.org/docs/for-app-authors/metainfo-guidelines#release

A description could be written for each release, but a URL for release
notes can be provided instead.

The specifications:
https://www.freedesktop.org/software/appstream/docs/

Flathub's own quality guidelines:
https://docs.flathub.org/docs/for-app-authors/metainfo-guidelines/quality-guidelines